+++ README.MD
@@ -3,9 +3,9 @@
 ## Hvem, hvad og hvorfor?
 Dette repository indeholder det udviklingsmæssige arbejde hen imod en færdigkørende applikation, bestilt af **Center for Dansk og Integration.** Ønsket fra CDI kan kortfattet formuleres som: *en web-app, der skal gøre det lettere for indvandrere at lære det danske sprog på arbejdspladsen.*
 
-Software-produktet er en fusion af et samarbejde mellem multimediedesignstuderende fra *EASJ Køge*, som har stået for en del af frontend-udviklingen og datamatikerstuderende fra *EASJ Næstved*, som har stået for backend-udvikling etc.
+Software-produktet er en fusion af et samarbejde mellem multimediedesignstuderende fra *EASJ Køge*, som har stået for en del af frontend-udviklingen (dog kun lettere inspiration reel anvendt) og datamatikerstuderende fra *EASJ Næstved*, som har stået for backend-udvikling, yderligere front-end etc.
 
-De officielle Product Owners er *Kirsten og Susanne* fra CDI, og web-app'en udgør en del af datamatikereksamen på tredje semester i Systemudvikling.
+De officielle Product Owners er *Kirsten og Susanne* fra CDI, og web-app'en udgør en del af datamatikereksamen på tredje semester i Systemudvikling, og indeholder desuden elementer, der indgår i eksamen for Programmering.
 
 I projektets Wiki-sektion præsenteres elementer fra systemudviklingsprocessen i form af eksempelvis Scrum-Board og Product Backlog.
