+++ package.json
@@ -7,11 +7,15 @@
   },
   "dependencies": {
     "bcryptjs": "^2.4.3",
+    "body-parser": "^1.18.3",
+    "connect-mongo": "^2.0.1",
     "cookie-parser": "~1.4.3",
     "debug": "~2.6.9",
     "express": "~4.16.0",
+    "express-session": "^1.15.6",
     "hbs": "~4.0.1",
     "http-errors": "~1.6.2",
+    "mongodb": "^3.1.8",
     "mongoose": "^5.3.8",
     "morgan": "~1.9.0"
   }
