+++ views/index.hbs
@@ -7,7 +7,8 @@
         <label for="inputEmail" class="sr-only">Email addresse</label>
         <input type="text" id="inputEmail" name="logemail" class="form-control" placeholder="Email addresse" required="">
         <label for="inputPassword" class="sr-only">Password</label>
-        <input type="password" name="logpassword" id="inputPassword" class="form-control" placeholder="Kodeord" required="">
+        <input type="password" name="logpassword" id="inputPassword" class="form-control" placeholder="Kodeord"
+            required="">
         <input class="btn btn-lg btn-primary btn-block" type="submit" value="LOGIN NU"></input>
         <p class="mt-5 mb-3 text-muted">&copy; 2017-2018</p>
     </form>
